Move tests into new namespace
