+++ tests/Presentation.Tests/Client/Pages/Admin/AdminPageTests.cs
@@ -3,7 +3,7 @@ using System.Net.Http;
 using System.Threading.Tasks;
 
 using Client.Pages.Admin;
-using Client.Pages.Contract;
+using Client.Pages.Contracts;
 
 using Domain.Contracts;
